moved CDF class into _cdf.py
